[Doc] Add docs about OpenAI compatible server (#3288)
