+++ docs/source/dev/sampling_params.rst
@@ -1,0 +1,4 @@
+Sampling Params
+===============
+
+.. automodule:: vllm.sampling_params.SamplingParams
